Canopus: usbd_hal.c -Delete force ZLP transmission for umass FIFO mode.
Delete the automatic ZLP transmission when the total length is a multiply
of endpoint size for umass ezamples.
If ZLP is necessary, please finish it in the driver rather than hal such
as usb_cdc example.
Issue description:
In umass initial process, a CSW packet is needed after CBW(read10) and 512
IN packet, but the device returns a ZLP which result in missing the status
and re-configure all the time.

Signed-off-by: LvTingting <Tingting.Lv@atmel.com>