Canopus: usbd_hal.c -Fix the data lose issue when TSC printer works on MAC
It is case about SAMA5D3, the number is 0026549.
Canopus has the same problem.
Description:
When printing a doc on Mac for the first time, no data lost.
Then print the sencond doc, the first 71 bytes will be lost on the
terminal but exist on the USB line.
DATA0/DATA1 data toggle is the root cause.
More details refer to salseforce and the emails.

Solution:
The following words are from usb spec.
 For endpoints using data toggle, regardless of whether an endpoint has the
 Halt feature set, a ClearFeature(ENDPOINT_HALT) request always results in the data toggle being reinitialized to DATA0.

Signed-off-by: LvTingting <Tingting.Lv@atmel.com>